requirements.txt
# Weather Dashboard - Production Dependencies
# ==========================================
# Pin all versions for reproducible builds
# Last updated: 2024

# =============================================================================
# CORE DEPENDENCIES
# =============================================================================

# HTTP and API Communication
requests==2.31.0              # HTTP client for weather API calls
aiohttp==3.9.1                # Async HTTP client for concurrent requests
httpx==0.25.2                 # Modern HTTP client with async support

# Configuration and Environment
python-dotenv==1.0.0          # Environment variable management
pydantic==2.5.2               # Data validation and settings management
pyyaml==6.0.1                 # YAML configuration file parsing
typing-extensions==4.8.0      # Extended type hints for better code quality

# =============================================================================
# DATABASE AND STORAGE
# =============================================================================

# Database Operations
sqlalchemy==2.0.23            # SQL toolkit and ORM
aiosqlite==0.19.0             # Async SQLite database operations
sqlite3                        # Built-in SQLite support (Python standard library)

# Caching and Performance
redis==5.0.1                  # Redis for advanced caching (optional)
cachetools==5.3.2             # In-memory caching utilities
ratelimit==2.2.1              # API rate limiting

# =============================================================================
# DATA PROCESSING AND ANALYSIS
# =============================================================================

# Numerical Computing
numpy==1.24.4                 # Fundamental package for scientific computing
pandas==2.1.4                 # Data manipulation and analysis
scipy==1.11.4                 # Scientific computing library

# Data Serialization
dataclasses-json==0.6.3       # JSON serialization for dataclasses
orjson==3.9.10                # Fast JSON library

# =============================================================================
# VISUALIZATION AND CHARTS
# =============================================================================

# Core Plotting
matplotlib==3.8.2             # Comprehensive plotting library
seaborn==0.13.0               # Statistical data visualization
plotly==5.17.0                # Interactive plotting library

# Image Processing
Pillow==10.1.0                # Python Imaging Library (PIL Fork)

# Document Generation
reportlab==4.0.7              # PDF generation library
wordcloud==1.9.2              # Word cloud generation

# =============================================================================
# USER INTERFACE
# =============================================================================

# Modern Tkinter Frameworks
customtkinter==5.2.2          # Modern and customizable tkinter
ttkbootstrap==1.10.1          # Bootstrap-themed tkinter widgets
tkinter                        # Built-in GUI toolkit (Python standard library)

# Web Framework (Optional)
streamlit==1.29.0             # Web app framework for data applications

# =============================================================================
# GEOLOCATION AND MAPPING
# =============================================================================

# Location Services
geopy==2.4.1                  # Geocoding library
geocoder==1.38.1              # Alternative geocoding service
timezonefinder==6.2.0         # Timezone lookup by coordinates
pytz==2023.3                  # World timezone definitions
folium==0.15.0                # Interactive maps
tkintermapview==1.29.0        # Interactive map widget for tkinter
tkinterweb==3.24.7            # Web browser widget for tkinter
pywebview==4.4.1              # Cross-platform webview with JavaScript support
branca==0.6.0                 # HTML/JS generation for folium
jinja2>=2.9                   # Template engine for folium

# =============================================================================
# AI AND MACHINE LEARNING
# =============================================================================

# AI API Clients
openai==1.6.1                 # OpenAI API client
google-generativeai==0.3.2    # Google Gemini AI API client

# Machine Learning (Optional)
scikit-learn==1.3.2           # Machine learning library
joblib==1.3.2                 # Lightweight pipelining and model persistence

# =============================================================================
# SYSTEM AND UTILITIES
# =============================================================================

# System Information
psutil==5.9.6                 # System and process utilities
platform                       # Access to platform-specific data (built-in)
os                             # Operating system interface (built-in)
sys                            # System-specific parameters (built-in)

# Date and Time
datetime                       # Date and time handling (built-in)
time                           # Time-related functions (built-in)

# Logging and Monitoring
loguru==0.7.2                 # Enhanced logging library
rich==13.7.0                  # Rich text and beautiful formatting

# =============================================================================
# ASYNC AND CONCURRENCY
# =============================================================================

asyncio                        # Asynchronous I/O (built-in)
concurrent.futures             # High-level interface for asynchronously executing callables (built-in)
threading                      # Thread-based parallelism (built-in)

# =============================================================================
# TESTING AND VALIDATION
# =============================================================================

# Data Validation
cerberus==1.3.5               # Lightweight data validation library
validators==0.22.0            # Python data validation library

# =============================================================================
# DEVELOPMENT DEPENDENCIES (Optional)
# =============================================================================
# Uncomment for development environment

# Code Quality
# black==23.12.1               # Code formatter
# isort==5.13.2                # Import sorter
# flake8==6.1.0                # Code linter
# mypy==1.8.0                  # Static type checker
# autoflake==2.2.1             # Remove unused imports and variables

# Testing
# pytest==7.4.3               # Testing framework
# pytest-asyncio==0.21.1      # Async testing support
# pytest-cov==4.1.0           # Coverage reporting
# pytest-mock==3.12.0         # Mocking library

# Documentation
# mkdocs==1.5.3                # Documentation generator
# mkdocs-material==9.5.3       # Material theme for MkDocs

# =============================================================================
# NOTES
# =============================================================================
# 
# 1. Built-in modules (tkinter, sqlite3, asyncio, etc.) don't need installation
# 2. Development dependencies are commented out - uncomment as needed
# 3. All versions are pinned for reproducible builds
# 4. Optional dependencies (Redis, ML libraries) can be removed if not used
# 5. For development: pip install -r requirements-dev.txt
# 6. For testing: pip install -r requirements-test.txt
#
